requirements.txt
asgiref==3.4.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==3.2.13
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
